[Please review & test: Feature]
I've implemented checkboxes for handlingFlags and modelFlags, seems to work fine in some situations, but some selections will disable other selections for some reason. I don't know the cause of this, nor how to fix it. 
Yes, the GUI is kinda simple. I'll improve it in the future, but I'd like to have working functions first.

The things to test: try if the selections have the correct effect, and the effects on different types of vehicles. Please note that modelFlags are buggy anyways, some values can't be set on some vehicles. Seems to depend on available dummies or something.

And the code is a little crappy, if anyone has ideas for a better code, please let me know :)

Thnx in advance.